Cast ActionController::Parameters to a hash before constructing Event
Without this change, nested hashes (e.g. `payload`) are not correctly
preserved. Example:

  params = ActionController::Parameters.new(payload: {foo:1}).permit(payload:{})
  event = Z::W::E.new(params)
  serialized = Zaikio::Webhooks::EventSerializer.serialize(event)
  JSON.dump(serialized)
  #=> "{\"_aj_serialized\":\"Zaikio::Webhooks::EventSerializer\",\"payload\":\"{\\\"foo\\\"=>1}\"}"

I believe this happens because we've got an OpenStruct containing this
Parameters object, rather than a plain hash, and our JSON library
doesn't know how to safely convert it.

If we cast to a hash first, everything works as expected.